Default button type now is Outline drawSpacedText() proc added Solid::Button's label is now handled by Solid::Button::draw_label() Checkbox and Dropdown now handle correctly the deactivated status

+++ include/solidbase.h
@@ -31,7 +31,10 @@ namespace Solid {
         }
 
         /// Construct from ints
-        Measure(int w, int h) : Width(w), Height(h) {
+//        Measure(int w, int h) : Width(w), Height(h) {
+//        };
+
+        Measure(float w, float h) : Width(w), Height(h) {
         };
     };
